test: move the NVDA stubs into a shared, parameterised module
tests_gui/ needs the same NVDA stubs but real wxPython, and copying the
apparatus would guarantee drift. install(stub_wx=...) is the one switch.

Also makes globalPluginHandler.GlobalPlugin a real class. Subclassing a
MagicMock() instance does not raise -- it silently yields a MagicMock and
discards the class body, so a test against the GlobalPlugin would have
passed while exercising nothing.

Refs #65
